+++ rq/scripts/__init__.py
@@ -13,6 +13,24 @@ def add_standard_arguments(parser):
             help='The Redis database (default: 0)')
 
 
+def read_config_file(module):
+    """Reads all UPPERCASE variables defined in the given module file."""
+    settings = __import__(module, [], [], [], -1)
+    return {k: v for k, v in settings.__dict__.items() if k.upper() == k}
+
+
+def setup_default_arguments(args, settings):
+    """ Sets up args from settings or defaults """
+    if args.host is None:
+        args.host = settings.get('REDIS_HOST', 'localhost')
+    if args.port is None:
+        args.port = int(settings.get('REDIS_PORT', 6379))
+    else:
+        args.port = int(args.port)
+    if args.db is None:
+        args.db = settings.get('REDIS_DB', 0)
+
+
 def setup_redis(args):
     redis_conn = redis.Redis(host=args.host, port=args.port, db=args.db)
     use_connection(redis_conn)
